i installed ts and I have started moving the files around. I weill begin by converting each file one by one.
