+++ index.html
@@ -12,6 +12,6 @@
     <body>
         <canvas id="gameCanvas"></canvas>
         <!-- Include the main script as a module -->
-        <script type="module" src="main.js"></script>
+        <script type="module" src="./src/main.js"></script>
     </body>
 </html>
