+++ package.json
@@ -4,17 +4,26 @@
     "description": "This README serves as a quick reference for setting up a new project environment from scratch. It will help ensure consistency across all projects and save time.",
     "main": "script.js",
     "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "lint": "eslint .",
-        "format": "prettier --write ."
+        "build": "tsc",
+        "lint": "eslint 'src/**/*.{js,ts}'",
+        "lint:fix": "eslint 'src/**/*.{js,ts}' --fix",
+        "format": "prettier --write 'src/**/*.{js,ts,json}'"
     },
     "keywords": [],
     "author": "",
     "license": "ISC",
     "devDependencies": {
         "@eslint/js": "^9.14.0",
+        "@typescript-eslint/eslint-plugin": "^8.14.0",
+        "@typescript-eslint/parser": "^8.14.0",
         "eslint": "^9.14.0",
+        "eslint-config-prettier": "^9.1.0",
+        "eslint-plugin-prettier": "^5.2.1",
         "globals": "^15.12.0",
-        "prettier": "^3.3.3"
+        "prettier": "^3.3.3",
+        "typescript": "^5.6.3"
+    },
+    "dependencies": {
+        "excalibur": "^0.29.3"
     }
 }
