+++ src/gameModules/targetingStrategies.js
