+++ style.css
@@ -5,7 +5,8 @@
     box-sizing: border-box;
 }
 
-body, html {
+body,
+html {
     width: 100%;
     height: 100%;
     display: flex;
@@ -17,8 +18,8 @@ body, html {
 
 #gameCanvas {
     border: 1px solid black;
-    width: auto;        /* Use full viewport width */
-    height: 100vh;        /* Height auto to maintain aspect ratio */
-    max-height: 100vh;   /* Limit height to viewport height */
+    width: auto; /* Use full viewport width */
+    height: 100vh; /* Height auto to maintain aspect ratio */
+    max-height: 100vh; /* Limit height to viewport height */
     aspect-ratio: 1080 / 1920; /* Maintain 1080x1920 aspect ratio */
 }
