+++ tsconfig.json
@@ -1,0 +1,13 @@
+{
+    "compilerOptions": {
+        "allowJs": true,
+        "outDir": "dist",
+        "rootDir": "src",
+        "target": "es6", // Set the ECMAScript target version, adjust if needed
+        "module": "commonjs", // Common module system, or use "esnext" if preferred
+        "strict": true, // Enable strict type-checking options
+        "esModuleInterop": true // Enables compatibility with ES modules
+    },
+    "include": ["src/**/*"],
+    "exclude": ["node_modules", "dist"]
+}
